Adding support/test for dual dialogue blocks
This required switching hunks_to_elements from a map function to a fold function
so that I can check the values of previous elements to see if they're dual
dialogue. I also have to run the fold in reverse since fountain's dual dialogue
indicator "^" appears after the second dialogue block and not the first.